+++ python/setup.py
@@ -18,8 +18,9 @@ if sys.platform.startswith('win32'):
     os.system('bash ' + CURRENT_DIR + '/pre.sh ' + python_path + ' ' + version)
 
     cabess_module = Extension(name='abess._cabess',
-                              sources=[CURRENT_DIR + '/src/abess.cpp', CURRENT_DIR + '/src/List.cpp', CURRENT_DIR + '/src/utilities.cpp',
-                                       CURRENT_DIR + '/src/normalize.cpp', CURRENT_DIR + '/src/abess.i'],
+                              sources=[CURRENT_DIR + '/src/api.cpp', CURRENT_DIR + '/src/List.cpp', CURRENT_DIR + '/src/utilities.cpp',
+                                       CURRENT_DIR + '/src/normalize.cpp', CURRENT_DIR + '/src/pywrap.cpp',
+                                       CURRENT_DIR + '/src/pywrap.i'],
                               language='c++',
                               extra_compile_args=["-DNDEBUG", "-fopenmp", "-O2", "-Wall", "-mavx", "-mfma", "-march=native",
                                                   "-std=c++11", "-mtune=generic", "-D%s" % os_type, path1, path2],
@@ -34,8 +35,9 @@ elif sys.platform.startswith('darwin'):
     # print(eigen_path)
     # eigen_path = "/usr/local/include/eigen3/Eigen"
     cabess_module = Extension(name='abess._cabess',
-                              sources=[CURRENT_DIR + '/src/abess.cpp', CURRENT_DIR + '/src/List.cpp', CURRENT_DIR + '/src/utilities.cpp',
-                                       CURRENT_DIR + '/src/normalize.cpp', CURRENT_DIR + '/src/abess.i'],
+                              sources=[CURRENT_DIR + '/src/api.cpp', CURRENT_DIR + '/src/List.cpp', CURRENT_DIR + '/src/utilities.cpp',
+                                       CURRENT_DIR + '/src/normalize.cpp', CURRENT_DIR + '/src/pywrap.cpp',
+                                       CURRENT_DIR + '/src/pywrap.i'],
                               language='c++',
                               extra_compile_args=[
                                   "-DNDEBUG", "-O2", "-Wall", "-std=c++11", "-mavx", "-mfma", "-march=native"],
@@ -47,8 +49,9 @@ else:
     # print(eigen_path)
     # eigen_path = "/usr/local/include/eigen3/Eigen"
     cabess_module = Extension(name='abess._cabess',
-                              sources=[CURRENT_DIR + '/src/abess.cpp', CURRENT_DIR + '/src/List.cpp', CURRENT_DIR + '/src/utilities.cpp',
-                                       CURRENT_DIR + '/src/normalize.cpp', CURRENT_DIR + '/src/abess.i'],
+                              sources=[CURRENT_DIR + '/src/api.cpp', CURRENT_DIR + '/src/List.cpp', CURRENT_DIR + '/src/utilities.cpp',
+                                       CURRENT_DIR + '/src/normalize.cpp', CURRENT_DIR + '/src/pywrap.cpp',
+                                       CURRENT_DIR + '/src/pywrap.i'],
                               language='c++',
                               extra_compile_args=[
                                   "-DNDEBUG", "-fopenmp", "-O2", "-Wall", "-std=c++11", "-mavx", "-mfma", "-march=native"],
