+++ python/src/abess.h
@@ -1,166 +1,43 @@
-/*****************************************************************************
-*  OpenST Basic tool library                                                 *
-*  Copyright (C) 2021 Kangkang Jiang  jiangkk3@mail2.sysu.edu.cn                         *
-*                                                                            *
-*  This file is part of OST.                                                 *
-*                                                                            *
-*  This program is free software; you can redistribute it and/or modify      *
-*  it under the terms of the GNU General Public License version 3 as         *
-*  published by the Free Software Foundation.                                *
-*                                                                            *
-*  You should have received a copy of the GNU General Public License         *
-*  along with OST. If not, see <http://www.gnu.org/licenses/>.               *
-*                                                                            *
-*  Unless required by applicable law or agreed to in writing, software       *
-*  distributed under the License is distributed on an "AS IS" BASIS,         *
-*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  *
-*  See the License for the specific language governing permissions and       *
-*  limitations under the License.                                            *
-*                                                                            *
-*  @file     abess.h                                                         *
-*  @brief    The main function of abess fremework                            *
-*                                                                            *
-*                                                                            *
-*  @author   Kangkang Jiang                                                  *
-*  @email    jiangkk3@mail2.sysu.edu.cn                                      *
-*  @version  0.0.1                                                           *
-*  @date     2021-07-31                                                      *
-*  @license  GNU General Public License (GPL)                                *
-*                                                                            *
-*----------------------------------------------------------------------------*
-*  Remark         : Description                                              *
-*----------------------------------------------------------------------------*
-*  Change History :                                                          *
-*  <Date>     | <Version> | <Author>       | <Description>                   *
-*----------------------------------------------------------------------------*
-*  2021/07/31 | 0.0.1     | Kangkang Jiang | First version                   *
-*----------------------------------------------------------------------------*
-*                                                                            *
-*****************************************************************************/
+#ifndef SRC_ABESS_H
+#define SRC_ABESS_H
 
-#ifndef BESS_BESS_H
-#define BESS_BESS_H
-
+// #define R_BUILD
 #ifdef R_BUILD
+
 #include <Rcpp.h>
 #include <RcppEigen.h>
 //[[Rcpp::depends(RcppEigen)]]
 using namespace Rcpp;
+
 #else
+
 #include <Eigen/Eigen>
 #include "List.h"
-#include "Algorithm.h"
+
 #endif
 
 #include <iostream>
+#include "Data.h"
+#include "Algorithm.h"
+#include "Metric.h"
+#include "path.h"
+#include "utilities.h"
+#include "screening.h"
+#include <vector>
 
-/** Result struct
- * @brief Save the sequential fitting result along the parameter searching.
- */
-template <class T2, class T3>
-struct Result
-{
-    Eigen::Matrix<T2, Eigen::Dynamic, Eigen::Dynamic> beta_matrix;            /*!<  */
-    Eigen::Matrix<T3, Eigen::Dynamic, Eigen::Dynamic> coef0_matrix;           /*!<  */
-    Eigen::MatrixXd ic_matrix;                                                /*!<  */
-    Eigen::MatrixXd test_loss_matrix;                                         /*!<  */
-    Eigen::MatrixXd train_loss_matrix;                                        /*!<  */
-    Eigen::Matrix<Eigen::VectorXd, Eigen::Dynamic, Eigen::Dynamic> bd_matrix; /*!<  */
-    Eigen::MatrixXd effective_number_matrix;                                  /*!<  */
-};
+typedef Eigen::Triplet<double> triplet;
 
-/** 
- * @brief The main function of abess fremework
- * @param X                             Training data.
- * @param y                             Target values. Will be cast to X's dtype if necessary.
- *                                      For linear regression problem, y should be a n time 1 numpy array with type \code{double}.
- *                                      For classification problem, \code{y} should be a $n \time 1$ numpy array with values \code{0} or \code{1}.
- *                                      For count data, \code{y} should be a $n \time 1$ numpy array of non-negative integer.
- * @param n                             Sample size.
- * @param p                             Variable dimension.
- * @param weight                        Individual weights for each sample. Only used for is_weight=True.
- * @param sigma                         Sample covariance matrix.For PCA, it can be given as input, instead of X. But if X is given, Sigma will be set to \code{np.cov(X.T)}.
- * @param is_normal                     Whether normalize the variables array before fitting the algorithm.
- * @param algorithm_type                Algorithm type.
- * @param model_type                    Model type.
- * @param max_iter                      Maximum number of iterations taken for the splicing algorithm to converge.
- *                                      Due to the limitation of loss reduction, the splicing algorithm must be able to converge.
- *                                      The number of iterations is only to simplify the implementation.
- * @param exchange_num                  Max exchange variable num.
- * @param path_type                     The method to be used to select the optimal support size. 
- *                                      For path_type = 1, we solve the best subset selection problem for each size in support_size. 
- *                                      For path_type = 2, we solve the best subset selection problem with support size ranged in (s_min, s_max), where the specific support size to be considered is determined by golden section.
- * @param is_warm_start                 When tuning the optimal parameter combination, whether to use the last solution as a warm start to accelerate the iterative convergence of the splicing algorithm.
- * @param ic_type                       The type of criterion for choosing the support size. Available options are "gic", "ebic", "bic", "aic".
- * @param is_cv                         Use the Cross-validation method to choose the support size.
- * @param Kold                          The folds number when Use the Cross-validation method.
- * @param sequence                      An integer vector representing the alternative support sizes. Only used for path_type = "seq".
- * @param s_min                         The lower bound of golden-section-search for sparsity searching.
- * @param s_max                         The higher bound of golden-section-search for sparsity searching.
- * @param K_max                         The max search time of golden-section-search for sparsity searching.
- * @param epsilon                       The stop condition of golden-section-search for sparsity searching.
- * @param thread                        Max number of multithreads. If thread = 0, the program will use the maximum number supported by the device.       
- * @param screen_size                   Screen the variables first and use the chosen variables in abess process. If screen_size = -1, screening will not be used.
- *                                      The number of variables remaining after screening. It should be a non-negative number smaller than p.
- * @param g_index                       The group index for each variable.
- * @param always_select                 An array contains the indexes of variables we want to consider in the model.
- * @param primary_model_fit_max_iter    The maximal number of iteration in `primary_model_fit()` (in Algorithm.h). 
- * @param primary_model_fit_epsilon     The epsilon (threshold) of iteration in `primary_model_fit()` (in Algorithm.h). 
- * @param splicing_type                 The type of splicing in `fit()` (in Algorithm.h). 
- *                                      "0" for decreasing by half, "1" for decresing by one.
- * @param sub_search                    The number of inactive sets that are split when splicing. It should be positive integer.
- * @return result list.
- */
-List abessGLM_API(Eigen::MatrixXd x, Eigen::MatrixXd y, int n, int p, int normalize_type,
-               Eigen::VectorXd weight, 
-               int algorithm_type, int model_type, int max_iter, int exchange_num,
-               int path_type, bool is_warm_start,
-               int ic_type, double ic_coef, int Kfold,
-               Eigen::VectorXi sequence,
-               Eigen::VectorXd lambda_seq,
-               int s_min, int s_max, 
-               double lambda_min, double lambda_max, int nlambda,
-               int screening_size, 
-               Eigen::VectorXi g_index,
-               Eigen::VectorXi always_select,
-               int primary_model_fit_max_iter, double primary_model_fit_epsilon,
-               bool early_stop, bool approximate_Newton,
-               int thread,
-               bool covariance_update,
-               bool sparse_matrix,
-               int splicing_type,
-               int sub_search,
-               Eigen::VectorXi cv_fold_id);
+using namespace Eigen;
+using namespace std;
 
-List abessPCA_API(Eigen::MatrixXd x,
-                  int n,
-                  int p,
-                  int normalize_type,
-                  Eigen::VectorXd weight,
-                  Eigen::MatrixXd sigma,
-                  int max_iter,
-                  int exchange_num,
-                  int path_type,
-                  bool is_warm_start,
-                  bool is_tune,
-                  int ic_type,
-                  double ic_coef,
-                  int Kfold,
-                  Eigen::VectorXi sequence,
-                  int s_min,
-                  int s_max,
-                  double lambda_min,
-                  double lambda_max,
-                  int nlambda,
-                  Eigen::VectorXi g_index,
-                  Eigen::VectorXi always_select,
-                  bool early_stop,
-                  int thread,
-                  bool sparse_matrix,
-                  int splicing_type,
-                  Eigen::VectorXi cv_fold_id,
-                  int pca_num);
-
+//  T1 for y, XTy, XTone
+//  T2 for beta
+//  T3 for coef0
+//  T4 for X
+//  <Eigen::VectorXd, Eigen::VectorXd, double, Eigen::MatrixXd> for Univariate Dense
+//  <Eigen::VectorXd, Eigen::VectorXd, double, Eigen::SparseMatrix<double> > for Univariate Sparse
+//  <Eigen::MatrixXd, Eigen::MatrixXd, Eigen::VectorXd, Eigen::MatrixXd> for Multivariable Dense
+//  <Eigen::MatrixXd, Eigen::MatrixXd, Eigen::VectorXd, Eigen::SparseMatrix<double> > for Multivariable Sparse
 template <class T1, class T2, class T3, class T4>
 List abessCpp(T4 &x, T1 &y, int n, int p, int normalize_type,
               Eigen::VectorXd weight, 
@@ -178,33 +55,231 @@ List abessCpp(T4 &x, T1 &y, int n, int p, int normalize_type,
               int thread,
               bool sparse_matrix,
               Eigen::VectorXi &cv_fold_id,
-              vector<Algorithm<T1, T2, T3, T4> *> algorithm_list);
-
+              vector<Algorithm<T1, T2, T3, T4> *> algorithm_list)
+{
 #ifndef R_BUILD
-void pywrap_abess(double *x, int x_row, int x_col, double *y, int y_row, int n, int p, int normalize_type, int y_col, double *weight, int weight_len, double *sigma, int sigma_row, int sigma_col,
-                  int algorithm_type, int model_type, int max_iter, int exchange_num,
-                  int path_type, bool is_warm_start,
-                  int ic_type, double ic_coef, int K,
-                  int *gindex, int gindex_len,
-                  int *sequence, int sequence_len,
-                  double *lambda_sequence, int lambda_sequence_len,
-                  int *cv_fold_id, int cv_fold_id_len,
-                  int s_min, int s_max, 
-                  double lambda_min, double lambda_max, int n_lambda,
-                  int screening_size, 
-                  int *always_select, int always_select_len, 
-                  int primary_model_fit_max_iter, double primary_model_fit_epsilon,
-                  bool early_stop, bool approximate_Newton,
-                  int thread,
-                  bool covariance_update,
-                  bool sparse_matrix,
-                  int splicing_type,
-                  int sub_search,
-                  int pca_num,
-                  double *beta_out, int beta_out_len, double *coef0_out, int coef0_out_len, double *train_loss_out,
-                  int train_loss_out_len, double *ic_out, int ic_out_len, double *nullloss_out, double *aic_out,
-                  int aic_out_len, double *bic_out, int bic_out_len, double *gic_out, int gic_out_len, int *A_out,
-                  int A_out_len, int *l_out);
+  std::srand(123);
 #endif
 
-#endif //BESS_BESS_H
+  // int normalize_type;
+  // switch (model_type)
+  // {
+  //     case 1: // gauss
+  //     case 5: // mul-gauss
+  //     case 7: // pca
+  //         normalize_type = 1; break;
+  //     case 2: // logi
+  //     case 3: // poiss
+  //     case 6: // mul-nomial
+  //         normalize_type = 2; break;
+  //     case 4: // cox
+  //         normalize_type = 3; break;
+  // };
+  
+  int algorithm_list_size = algorithm_list.size();
+  int beta_size = algorithm_list[0]->get_beta_size(n, p); // number of candidate param
+
+  // data packing
+  Data<T1, T2, T3, T4> data(x, y, normalize_type, weight, g_index, sparse_matrix, beta_size);
+
+  // screening
+  Eigen::VectorXi screening_A; 
+  if (screening_size >= 0) 
+  {
+    screening_A = screening<T1, T2, T3, T4>(data, algorithm_list, screening_size, beta_size, lambda_seq[0]);
+  }
+
+  // For CV:
+  // 1:mask
+  // 2:warm start save
+  // 3:group_XTX
+  Metric<T1, T2, T3, T4> *metric = new Metric<T1, T2, T3, T4>(ic_type, ic_coef, Kfold);
+  if (Kfold > 1){
+    metric->set_cv_train_test_mask(data, data.n, cv_fold_id);
+    metric->set_cv_init_fit_arg(beta_size, data.M);
+    // metric->set_cv_initial_model_param(Kfold, data.p);
+    // metric->set_cv_initial_A(Kfold, data.p);
+    // metric->set_cv_initial_coef0(Kfold, data.p);
+    // if (model_type == 1)
+    //   metric->cal_cv_group_XTX(data);
+  }
+
+  // calculate loss for each parameter parameter combination
+  vector<Result<T2, T3>> result_list(Kfold);
+  if (path_type == 1)
+  {
+#pragma omp parallel for
+    for (int i = 0; i < Kfold; i++)
+    {
+      sequential_path_cv<T1, T2, T3, T4>(data, algorithm_list[i], metric, sequence, lambda_seq, early_stop, i, result_list[i]);
+    }
+  }
+  else
+  {
+    // if (algorithm_type == 5 || algorithm_type == 3)
+    // {
+    //     double log_lambda_min = log(max(lambda_min, 1e-5));
+    //     double log_lambda_max = log(max(lambda_max, 1e-5));
+
+    //     result = pgs_path(data, algorithm, metric, s_min, s_max, log_lambda_min, log_lambda_max, powell_path, nlambda);
+    // }
+    gs_path<T1, T2, T3, T4>(data, algorithm_list, metric, s_min, s_max, sequence, lambda_seq, result_list);
+  }
+
+  for (int k = 0; k < Kfold; k++)
+  {
+    algorithm_list[k]->clear_setting();
+  }
+
+  // Get bestmodel index && fit bestmodel
+  int min_loss_index_row = 0, min_loss_index_col = 0, s_size = sequence.size(), lambda_size = lambda_seq.size();
+  Eigen::Matrix<T2, Dynamic, Dynamic> beta_matrix(s_size, lambda_size);
+  Eigen::Matrix<T3, Dynamic, Dynamic> coef0_matrix(s_size, lambda_size);
+  Eigen::Matrix<VectorXd, Dynamic, Dynamic> bd_matrix(s_size, lambda_size);
+  Eigen::MatrixXd ic_matrix(s_size, lambda_size);
+  Eigen::MatrixXd test_loss_sum = Eigen::MatrixXd::Zero(s_size, lambda_size);
+  Eigen::MatrixXd train_loss_matrix(s_size, lambda_size);
+  Eigen::MatrixXd effective_number_matrix(s_size, lambda_size);
+
+  if (Kfold == 1) 
+  {
+    beta_matrix = result_list[0].beta_matrix;
+    coef0_matrix = result_list[0].coef0_matrix;
+    ic_matrix = result_list[0].ic_matrix;
+    train_loss_matrix = result_list[0].train_loss_matrix;
+    effective_number_matrix = result_list[0].effective_number_matrix;
+    ic_matrix.minCoeff(&min_loss_index_row, &min_loss_index_col);
+  } 
+  else 
+  { 
+    Eigen::MatrixXd test_loss_tmp;
+    for (int i = 0; i < Kfold; i++)
+    {
+      test_loss_tmp = result_list[i].test_loss_matrix;
+      test_loss_sum = test_loss_sum + test_loss_tmp / Kfold;
+    }
+    test_loss_sum.minCoeff(&min_loss_index_row, &min_loss_index_col);
+
+    Eigen::VectorXi used_algorithm_index = Eigen::VectorXi::Zero(algorithm_list_size);
+
+    // refit on full data
+#pragma omp parallel for
+    for (int i = 0; i < sequence.size() * lambda_seq.size(); i++)
+    {
+      int s_index = i / lambda_seq.size();
+      int lambda_index = i % lambda_seq.size();
+      int algorithm_index = omp_get_thread_num();
+      used_algorithm_index(algorithm_index) = 1;
+
+      T2 beta_init;
+      T3 coef0_init;
+      coef_set_zero(beta_size, data.M, beta_init, coef0_init);
+      Eigen::VectorXd bd_init = Eigen::VectorXd::Zero(data.g_num);
+
+      // warmstart from CV's result
+      for (int j = 0; j < Kfold; j++)
+      {
+        beta_init = beta_init + result_list[j].beta_matrix(s_index, lambda_index) / Kfold;
+        coef0_init = coef0_init + result_list[j].coef0_matrix(s_index, lambda_index) / Kfold;
+        bd_init = bd_init + result_list[j].bd_matrix(s_index, lambda_index) / Kfold;
+      }
+
+      algorithm_list[algorithm_index]->update_sparsity_level(sequence(s_index));
+      algorithm_list[algorithm_index]->update_lambda_level(lambda_seq(lambda_index));
+      algorithm_list[algorithm_index]->update_beta_init(beta_init);
+      algorithm_list[algorithm_index]->update_coef0_init(coef0_init);
+      algorithm_list[algorithm_index]->update_bd_init(bd_init);
+      algorithm_list[algorithm_index]->fit(data.x, data.y, data.weight, data.g_index, data.g_size, data.n, data.p, data.g_num);
+
+      beta_matrix(s_index, lambda_index) = algorithm_list[algorithm_index]->get_beta();
+      coef0_matrix(s_index, lambda_index) = algorithm_list[algorithm_index]->get_coef0();
+      train_loss_matrix(s_index, lambda_index) = algorithm_list[algorithm_index]->get_train_loss();
+      ic_matrix(s_index, lambda_index) = metric->ic(data.n, data.M, data.g_num, algorithm_list[algorithm_index]);
+      effective_number_matrix(s_index, lambda_index) = algorithm_list[algorithm_index]->get_effective_number();
+    }
+
+    for (int i = 0; i < algorithm_list_size; i++)
+      if (used_algorithm_index(i) == 1)
+      {
+        algorithm_list[i]->clear_setting();
+      }
+  }
+
+  // best_fit_result (output)
+  // int best_s = sequence(min_loss_index_row);
+  double best_lambda = lambda_seq(min_loss_index_col);
+
+  T2 best_beta;
+  T3 best_coef0;
+  double best_train_loss, best_ic, best_test_loss;
+
+  best_beta = beta_matrix(min_loss_index_row, min_loss_index_col);
+  best_coef0 = coef0_matrix(min_loss_index_row, min_loss_index_col);
+  best_train_loss = train_loss_matrix(min_loss_index_row, min_loss_index_col);
+  best_ic = ic_matrix(min_loss_index_row, min_loss_index_col);
+  best_test_loss = test_loss_sum(min_loss_index_row, min_loss_index_col);
+
+  //////////////Restore best_fit_result for normal//////////////
+  restore_for_normal<T2, T3>(best_beta, best_coef0, beta_matrix, coef0_matrix, sparse_matrix,
+                             data.normalize_type, data.n, data.x_mean, data.y_mean, data.x_norm);
+
+  // List result;
+  List out_result;
+#ifdef R_BUILD
+  out_result = List::create(Named("beta") = best_beta,
+                            Named("coef0") = best_coef0,
+                            Named("train_loss") = best_train_loss,
+                            Named("ic") = best_ic,
+                            Named("lambda") = best_lambda,
+                            Named("beta_all") = beta_matrix,
+                            Named("coef0_all") = coef0_matrix,
+                            Named("train_loss_all") = train_loss_matrix,
+                            Named("ic_all") = ic_matrix,
+                            Named("effective_number_all") = effective_number_matrix,
+                            Named("test_loss_all") = test_loss_sum);
+  if (path_type == 2)
+  {
+    out_result.push_back(sequence, "sequence");
+  }
+#else
+  out_result.add("beta", best_beta);
+  out_result.add("coef0", best_coef0);
+  out_result.add("train_loss", best_train_loss);
+  out_result.add("test_loss", best_test_loss);
+  out_result.add("ic", best_ic);
+  out_result.add("lambda", best_lambda);
+  // out_result.add("beta_all", beta_matrix);
+  // out_result.add("coef0_all", coef0_matrix);
+  // out_result.add("train_loss_all", train_loss_matrix);
+  // out_result.add("ic_all", ic_matrix);
+  // out_result.add("test_loss_all", test_loss_sum);
+#endif
+
+  // Restore best_fit_result for screening
+  if (screening_size >= 0)
+  {
+
+    T2 beta_screening_A;
+    T2 beta;
+    T3 coef0;
+    beta_size = algorithm_list[0]->get_beta_size(n, p);
+    coef_set_zero(beta_size, data.M, beta, coef0);
+
+#ifndef R_BUILD
+    out_result.get_value_by_name("beta", beta_screening_A);
+    slice_restore(beta_screening_A, screening_A, beta);
+    out_result.add("beta", beta);
+    out_result.add("screening_A", screening_A);
+#else
+    beta_screening_A = out_result["beta"];
+    slice_restore(beta_screening_A, screening_A, beta);
+    out_result["beta"] = beta;
+    out_result.push_back(screening_A, "screening_A");
+#endif
+  }
+
+  delete metric;
+  return out_result;
+}
+
+#endif // SRC_ABESS_H
