+++ python/src/normalize.h
@@ -2,8 +2,8 @@
 // Created by Jin Zhu on 2020/3/8.
 //
 // #define R_BUILD
-#ifndef ABESS_NORMALIZE_H
-#define ABESS_NORMALIZE_H
+#ifndef SRC_NORMALIZE_H
+#define SRC_NORMALIZE_H
 
 #ifdef R_BUILD
 #include <RcppEigen.h>
@@ -21,4 +21,4 @@ void Normalize(Eigen::SparseMatrix<double> &X, Eigen::MatrixXd &y, Eigen::Vector
 void Normalize3(Eigen::SparseMatrix<double> &X, Eigen::VectorXd &y, Eigen::VectorXd &meanx, Eigen::VectorXd &normx);
 void Normalize4(Eigen::SparseMatrix<double> &X, Eigen::VectorXd &y, Eigen::VectorXd &normx);
 
-#endif //BESS_NORMALIZE_H
+#endif // SRC_NORMALIZE_H
