+++ python/src/path.h
@@ -18,7 +18,6 @@ using namespace Eigen;
 #include "Data.h"
 #include "Algorithm.h"
 #include "Metric.h"
-#include "abess.h"
 #include "utilities.h"
 #include <vector>
